# NoticiasApi/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Noticias API MX",
    "version": "1.0.0",
    "description": "API de noticias con contribución y aprobación por admin. Formato de datos según 4.2."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "Users" },
    { "name": "Categories" },
    { "name": "States" },
    { "name": "News" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Perfil": { "type": "object", "properties": { "id": { "type": "integer" }, "nombre": { "type": "string" } } },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "profile_id": { "type": "integer", "enum": [1,2] },
          "nombres": { "type": "string" },
          "apellidos": { "type": "string" },
          "nick": { "type": "string" },
          "correo": { "type": "string", "format": "email" },
          "activo": { "type": "boolean" },
          "UserAlta": { "type": "string" }, "FechaAlta": { "type": "string", "format": "date-time" },
          "UserMod": { "type": "string" }, "FechaMod": { "type": "string", "format": "date-time" },
          "UserBaja": { "type": "string" }, "FechaBaja": { "type": "string", "format": "date-time" }
        }
      },
      "Estado": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nombre": { "type": "string" },
          "abreviacion": { "type": "string" },
          "activo": { "type": "boolean" },
          "UserAlta": { "type": "string" }, "FechaAlta": { "type": "string", "format": "date-time" },
          "UserMod": { "type": "string" }, "FechaMod": { "type": "string", "format": "date-time" },
          "UserBaja": { "type": "string" }, "FechaBaja": { "type": "string", "format": "date-time" }
        }
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nombre": { "type": "string" },
          "descripcion": { "type": "string" },
          "activo": { "type": "boolean" },
          "UserAlta": { "type": "string" }, "FechaAlta": { "type": "string", "format": "date-time" },
          "UserMod": { "type": "string" }, "FechaMod": { "type": "string", "format": "date-time" },
          "UserBaja": { "type": "string" }, "FechaBaja": { "type": "string", "format": "date-time" }
        }
      },
      "Noticia": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "categoria_id": { "type": "string", "format": "uuid" },
          "estado_id": { "type": "string", "format": "uuid" },
          "usuario_id": { "type": "string", "format": "uuid" },
          "titulo": { "type": "string" },
          "fecha_publicacion": { "type": "string", "format": "date" },
          "description": { "type": "string" },
          "image": { "type": "string", "description": "base64" },
          "estatus": { "type": "string", "enum": ["pendiente","aprobada","rechazada"] },
          "activo": { "type": "boolean" },
          "UserAlta": { "type": "string" }, "FechaAlta": { "type": "string", "format": "date-time" },
          "UserMod": { "type": "string" }, "FechaMod": { "type": "string", "format": "date-time" },
          "UserBaja": { "type": "string" }, "FechaBaja": { "type": "string", "format": "date-time" }
        }
      },
      "LoginInput": { "type": "object", "required": ["correo","contraseña"], "properties": { "correo": { "type": "string" }, "contraseña": { "type": "string" } } },
      "RegistroInput": { "type": "object", "required": ["nombres","apellidos","nick","correo","contraseña"], "properties": { "nombres": { "type": "string" }, "apellidos": { "type": "string" }, "nick": { "type": "string" }, "correo": { "type": "string" }, "contraseña": { "type": "string" } } },
      "Error": { "type": "object", "properties": { "error": { "type": "string" } } }
    }
  },
  "paths": {
    "/api/auth/registro": {
      "post": {
        "tags": ["Auth"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegistroInput" } } } },
        "responses": { "201": { "description": "Registrado" }, "409": { "description": "Correo duplicado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginInput" } } } },
        "responses": { "200": { "description": "OK" }, "401": { "description": "Credenciales inválidas" } }
      }
    },
    "/api/users": {
      "get": { "tags": ["Users"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Lista usuarios" } } },
      "post": { "tags": ["Users"], "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Creado" } } }
    },
    "/api/users/{id}": {
      "get": { "tags": ["Users"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" }, "404": { "description": "No existe" } } },
      "put": { "tags": ["Users"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Actualizado" } } },
      "delete": { "tags": ["Users"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Baja lógica" } } }
    },
    "/api/categories": {
      "get": { "tags": ["Categories"], "responses": { "200": { "description": "OK" } } },
      "post": { "tags": ["Categories"], "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Creada" } } }
    },
    "/api/categories/{id}": {
      "get": { "tags": ["Categories"], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" }, "404": { "description": "No existe" } } },
      "put": { "tags": ["Categories"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Actualizada" } } },
      "delete": { "tags": ["Categories"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Baja lógica" } } }
    },
    "/api/states": {
      "get": { "tags": ["States"], "responses": { "200": { "description": "OK" } } },
      "post": { "tags": ["States"], "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Creado" } } }
    },
    "/api/states/{id}": {
      "get": { "tags": ["States"], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } },
      "put": { "tags": ["States"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Actualizado" } } },
      "delete": { "tags": ["States"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Baja lógica" } } }
    },
    "/api/news": {
      "get": { "tags": ["News"], "parameters": [
        { "in": "query", "name": "categoria_id", "schema": { "type": "string" } },
        { "in": "query", "name": "estado_id", "schema": { "type": "string" } },
        { "in": "query", "name": "estatus", "schema": { "type": "string", "enum": ["pendiente","aprobada","rechazada"] } }
      ], "responses": { "200": { "description": "Listado público (aprobadas)" } } },
      "post": { "tags": ["News"], "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Creada (pendiente)" } } }
    },
    "/api/news/{id}": {
      "get": { "tags": ["News"], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Noticia aprobada" }, "403": { "description": "No disponible" }, "404": { "description": "No existe" } } },
      "put": { "tags": ["News"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Actualizada" } } },
      "delete": { "tags": ["News"], "security": [{ "bearerAuth": [] }], "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Baja lógica" } } }
    }
  }
}